Factor out portions of PendingClusterState bound to a specific bucket space to PendingBucketSpaceDbTransition.
