+++ storage/src/vespa/storage/distributor/pendingclusterstate.cpp
@@ -1,6 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "pendingclusterstate.h"
+#include "pending_bucket_space_db_transition.h"
 #include "bucketdbupdater.h"
 #include <vespa/storageframework/defaultimplementation/clock/realclock.h>
 #include <vespa/storage/common/bucketoperationlogger.h>
@@ -28,14 +29,14 @@ PendingClusterState::PendingClusterState(
     : _cmd(newStateCmd),
       _requestedNodes(newStateCmd->getSystemState().getNodeCount(lib::NodeType::STORAGE)),
       _outdatedNodes(newStateCmd->getSystemState().getNodeCount(lib::NodeType::STORAGE)),
-      _iter(0),
       _prevClusterState(clusterInfo->getClusterState()),
       _newClusterState(newStateCmd->getSystemState()),
       _clock(clock),
       _clusterInfo(clusterInfo),
       _creationTimestamp(creationTimestamp),
       _sender(sender),
-      _bucketOwnershipTransfer(distributorChanged(_prevClusterState, _newClusterState))
+      _bucketOwnershipTransfer(distributorChanged(_prevClusterState, _newClusterState)),
+      _pendingTransition()
 {
     logConstructionInformation();
     if (hasBucketOwnershipTransfer()) {
@@ -44,6 +45,7 @@ PendingClusterState::PendingClusterState(
         updateSetOfNodesThatAreOutdated();
         addAdditionalNodesToOutdatedSet(outdatedNodes);
     }
+    _pendingTransition = std::make_unique<PendingBucketSpaceDbTransition>(*this, _clusterInfo, _newClusterState, _creationTimestamp);
     if (shouldRequestBucketInfo()) {
         requestNodes();
     }
@@ -56,17 +58,18 @@ PendingClusterState::PendingClusterState(
         api::Timestamp creationTimestamp)
     : _requestedNodes(clusterInfo->getStorageNodeCount()),
       _outdatedNodes(clusterInfo->getStorageNodeCount()),
-      _iter(0),
       _prevClusterState(clusterInfo->getClusterState()),
       _newClusterState(clusterInfo->getClusterState()),
       _clock(clock),
       _clusterInfo(clusterInfo),
       _creationTimestamp(creationTimestamp),
       _sender(sender),
-      _bucketOwnershipTransfer(true)
+      _bucketOwnershipTransfer(true),
+      _pendingTransition()
 {
     logConstructionInformation();
     markAllAvailableNodesAsRequiringRequest();
+    _pendingTransition = std::make_unique<PendingBucketSpaceDbTransition>(*this, _clusterInfo, _newClusterState, _creationTimestamp);
     if (shouldRequestBucketInfo()) {
         requestNodes();
     }
@@ -364,14 +367,7 @@ PendingClusterState::onRequestBucketInfoReply(const std::shared_ptr<api::Request
     }
 
     setNodeReplied(node);
-
-    for (uint32_t i = 0; i < reply->getBucketInfo().size(); ++i) {
-        addNodeInfo(reply->getBucketInfo()[i]._bucketId,
-                    BucketCopy(_creationTimestamp,
-                               node,
-                               reply->getBucketInfo()[i]._info));
-    }
-
+    _pendingTransition->onRequestBucketInfoReply(*reply, node);
     _sentMessages.erase(iter);
 
     return true;
@@ -389,68 +385,8 @@ PendingClusterState::resendDelayedMessages() {
     }
 }
 
-void
-PendingClusterState::addNodeInfo(
-        const document::BucketId& id,
-        const BucketCopy& copy)
-{
-    _entries.push_back(Entry(id, copy));
-}
-
-PendingClusterState::Range
-PendingClusterState::skipAllForSameBucket()
-{
-    Range r(_iter, _iter);
-
-    for (document::BucketId& bid = _entries[_iter].bucketId;
-         _iter < _entries.size() && _entries[_iter].bucketId == bid;
-         ++_iter)
-    {
-    }
-
-    r.second = _iter;
-    return r;
-}
-
-void
-PendingClusterState::insertInfo(
-        BucketDatabase::Entry& info,
-        const Range& range)
-{
-    std::vector<BucketCopy> copiesToAddOrUpdate(
-            getCopiesThatAreNewOrAltered(info, range));
-
-    std::vector<uint16_t> order(
-            _clusterInfo->getIdealStorageNodesForState(
-                    _newClusterState,
-                    _entries[range.first].bucketId));
-    info->addNodes(copiesToAddOrUpdate, order, TrustedUpdate::DEFER);
-
-    LOG_BUCKET_OPERATION_NO_LOCK(
-            _entries[range.first].bucketId,
-            vespalib::make_string("insertInfo: %s",
-                                        info.toString().c_str()));
-}
-
-std::vector<BucketCopy>
-PendingClusterState::getCopiesThatAreNewOrAltered(
-        BucketDatabase::Entry& info,
-        const Range& range)
-{
-    std::vector<BucketCopy> copiesToAdd;
-    for (uint32_t i = range.first; i < range.second; ++i) {
-        const BucketCopy& candidate(_entries[i].copy);
-        const BucketCopy* cp = info->getNode(candidate.getNode());
-
-        if (!cp || !(cp->getBucketInfo() == candidate.getBucketInfo())) {
-            copiesToAdd.push_back(candidate);
-        }
-    }
-    return copiesToAdd;
-}
-
 std::string
-PendingClusterState::requestNodesToString()
+PendingClusterState::requestNodesToString() const
 {
     std::ostringstream ost;
     for (uint32_t i = 0; i < _requestedNodes.size(); ++i) {
@@ -464,137 +400,10 @@ PendingClusterState::requestNodesToString()
     return ost.str();
 }
 
-bool
-PendingClusterState::removeCopiesFromNodesThatWereRequested(
-        BucketDatabase::Entry& e,
-        const document::BucketId& bucketId)
-{
-    bool updated = false;
-    for (uint32_t i = 0; i < e->getNodeCount();) {
-        auto& info(e->getNodeRef(i));
-        const uint16_t entryNode(info.getNode());
-        // Don't remove an entry if it's been updated in the time after the
-        // bucket info requests were sent, as this would erase newer state.
-        // Don't immediately update trusted state, as that could erroneously
-        // mark a single remaining replica as trusted even though there might
-        // be one or more additional replicas pending merge into the database.
-        if (nodeIsOutdated(entryNode)
-            && (info.getTimestamp() < _creationTimestamp)
-            && e->removeNode(entryNode, TrustedUpdate::DEFER))
-        {
-            LOG(spam,
-                "Removed bucket %s from node %d",
-                bucketId.toString().c_str(),
-                entryNode);
-            updated = true;
-            // After removing current node, getNodeRef(i) will point to the _next_ node, so don't increment `i`.
-        } else {
-            ++i;
-        }
-    }
-    return updated;
-}
-
-bool
-PendingClusterState::databaseIteratorHasPassedBucketInfoIterator(
-        const document::BucketId& bucketId) const
-{
-    return (_iter < _entries.size()
-            && _entries[_iter].bucketId.toKey() < bucketId.toKey());
-}
-
-bool
-PendingClusterState::bucketInfoIteratorPointsToBucket(
-        const document::BucketId& bucketId) const
-{
-    return _iter < _entries.size() && _entries[_iter].bucketId == bucketId;
-}
-
-bool
-PendingClusterState::process(BucketDatabase::Entry& e)
-{
-    document::BucketId bucketId(e.getBucketId());
-
-    LOG(spam,
-        "Before merging info from nodes [%s], bucket %s had info %s",
-        requestNodesToString().c_str(),
-        bucketId.toString().c_str(),
-        e.getBucketInfo().toString().c_str());
-
-    while (databaseIteratorHasPassedBucketInfoIterator(bucketId)) {
-        LOG(spam, "Found new bucket %s, adding",
-            _entries[_iter].bucketId.toString().c_str());
-
-        _missingEntries.push_back(skipAllForSameBucket());
-    }
-
-    bool updated(removeCopiesFromNodesThatWereRequested(e, bucketId));
-
-    if (bucketInfoIteratorPointsToBucket(bucketId)) {
-        LOG(spam, "Updating bucket %s",
-            _entries[_iter].bucketId.toString().c_str());
-
-        insertInfo(e, skipAllForSameBucket());
-        updated = true;
-    }
-
-    if (updated) {
-        // Remove bucket if we've previously removed all nodes from it
-        if (e->getNodeCount() == 0) {
-            _removedBuckets.push_back(bucketId);
-        } else {
-            e.getBucketInfo().updateTrusted();
-        }
-    }
-
-    LOG(spam,
-        "After merging info from nodes [%s], bucket %s had info %s",
-        requestNodesToString().c_str(),
-        bucketId.toString().c_str(),
-        e.getBucketInfo().toString().c_str());
-
-    return true;
-}
-
-void
-PendingClusterState::addToBucketDB(BucketDatabase& db,
-                                   const Range& range)
-{
-    LOG(spam, "Adding new bucket %s with %d copies",
-        _entries[range.first].bucketId.toString().c_str(),
-        range.second - range.first);
-
-    BucketDatabase::Entry e(_entries[range.first].bucketId, BucketInfo());
-    insertInfo(e, range);
-    if (e->getLastGarbageCollectionTime() == 0) {
-        e->setLastGarbageCollectionTime(
-                framework::MicroSecTime(_creationTimestamp)
-                    .getSeconds().getTime());
-    }
-    e.getBucketInfo().updateTrusted();
-    db.update(e);
-}
-
 void
 PendingClusterState::mergeInto(BucketDatabase& db)
 {
-    std::sort(_entries.begin(), _entries.end());
-
-    db.forEach(*this);
-
-    for (uint32_t i = 0; i < _removedBuckets.size(); ++i) {
-        db.remove(_removedBuckets[i]);
-    }
-    _removedBuckets.clear();
-
-    // All of the remaining were not already in the bucket database.
-    while (_iter < _entries.size()) {
-        _missingEntries.push_back(skipAllForSameBucket());
-    }
-
-    for (uint32_t i = 0; i < _missingEntries.size(); ++i) {
-        addToBucketDB(db, _missingEntries[i]);
-    }
+    _pendingTransition->mergeInto(db);
 }
 
 void
@@ -619,6 +428,18 @@ PendingClusterState::getSummary() const
     return Summary(_prevClusterState.toString(),
                    _newClusterState.toString(),
                    (_clock.getTimeInMicros().getTime() - _creationTimestamp));
+}
+
+const PendingBucketSpaceDbTransition::EntryList &
+PendingClusterState::results() const
+{
+    return _pendingTransition->results();
+}
+
+void
+PendingClusterState::addNodeInfo(const document::BucketId& id, const BucketCopy& copy)
+{
+    _pendingTransition->addNodeInfo(id, copy);
 }
 
 }
